Call the notification email method regardless of whether an email is set
Because the notification method can be overridden by a user, it
should be called whether or not the account_verify_notification_email
is set in the adapter. In the adapter method, we don't need to check
if account_verify_notification_email is set, because the send_mail
command does nothing if the recipient is None.