Check workspace ACLs if workspace access is NO ACCESS for app workspaces
When checking workspace ownership, we unfortunately can't rely on
the results in list_workspaces. This is because the API currently
returns "NO ACCESS" when the workspace is shared with the app as an
OWNER, but the service account is not part of the auth domain. The
workaround is to check the ACLs when access is listed as NO ACCESS.
If the API returns a 404, the app is not an owner.