Use a different API to check requester pays status
This API works even if we are an owner but with no access to a
workspace, while the get_workspace only works if you have access.
Add a new method to the AnVILAPIClient (get_workspace_settings) that
calls the new API method. Update workspace audit and tests to use
the new method instead of the method to list bucket options.