Show accessible and unknown workspaces in separate tables
Add a new table to show workspaces with unknown access for an
account with an indicator about what is unknown (sharing vs
auth domain). This table expects each workspace object to have
the sharing_known and auth_domain_known fields added to the
instances - it won't be able to pull them from the database.
Use the new table on the AccountDetail page to show the set of
workspaces with unknown access.