Allow NO ACCESS workspaces to be imported with WorkspaceImport
Populate the list of workspaces by selecting those with an access
level of NO ACCESS or OWNER. Add tests to check that a workspace
where we are owners but not members of the auth domain is properly
imported using the view.

NB: It is possible that we are not owners of a NO ACCESS workspace
(e.g., if it was shared with us as a READER but we are not in the
auth domain), but I don't think it's worth solving this issue at
this time. The method to actually import the workspace will fail.