Cache audit results in run_anvil_audit command
This required changing how the caching keys were obtained. Add a method
to the AuditRun classes to get the cache key for that instance. Keep
the method in the AuditReview mixin to get the cache key, since we
don't have an instance - we have to pull the instance from the cache.