Check that cache MAX_ENTRIES option is large enough
When running AnVIL audits with cache=True, check that the cache
MAX_ENTRIES option is large enough. This requires using a database
cache, so also check that the cache type is set correctly. Require
the user to specify a cache name in the settings file instead of
using "default" as the default.